style: fix signedness warnings in test with static_cast<size_t>

+++ lib/test_ringbuffer.cc
@@ -8,11 +8,11 @@ TEST(RingBuffer, BasicPutGet) {
 
   EXPECT_TRUE(ringbuffer_is_empty(&rb));
   EXPECT_FALSE(ringbuffer_is_full(&rb));
-  EXPECT_EQ(ringbuffer_size(&rb), static_cast<size_t>(0));
-  EXPECT_EQ(ringbuffer_capacity(&rb), static_cast<size_t>(3)); // 4 - 1
+  EXPECT_EQ(ringbuffer_size(&rb), static_cast<size_t>(0U));
+  EXPECT_EQ(ringbuffer_capacity(&rb), static_cast<size_t>(3U)); // 4 - 1
 
   EXPECT_EQ(ringbuffer_put(&rb, 'a'), 0);
-  EXPECT_EQ(ringbuffer_size(&rb), static_cast<size_t>(1));
+  EXPECT_EQ(ringbuffer_size(&rb), static_cast<size_t>(1U));
   EXPECT_FALSE(ringbuffer_is_empty(&rb));
 
   uint8_t ch;
@@ -31,7 +31,7 @@ TEST(RingBuffer, FullCondition) {
   EXPECT_EQ(ringbuffer_put(&rb, '3'), 0);
   
   EXPECT_TRUE(ringbuffer_is_full(&rb));
-  EXPECT_EQ(ringbuffer_size(&rb), static_cast<size_t>(3));
+  EXPECT_EQ(ringbuffer_size(&rb), static_cast<size_t>(3U));
   
   // This should fail
   EXPECT_EQ(ringbuffer_put(&rb, '4'), -1);
@@ -52,15 +52,15 @@ TEST(RingBuffer, WrapAround) {
   ringbuffer_get(&rb, &ch);
   
   // wr_pos should be at 2, rd_pos at 2
-  EXPECT_EQ(rb.wr_pos, 2);
-  EXPECT_EQ(rb.rd_pos, 2);
+  EXPECT_EQ(rb.wr_pos, static_cast<size_t>(2U));
+  EXPECT_EQ(rb.rd_pos, static_cast<size_t>(2U));
   
   // Put 2 more to trigger wrap around
   EXPECT_EQ(ringbuffer_put(&rb, '3'), 0); // wr_pos -> 3
   EXPECT_EQ(ringbuffer_put(&rb, '4'), 0); // wr_pos -> 0 (wrap)
   
-  EXPECT_EQ(rb.wr_pos, 0);
-  EXPECT_EQ(ringbuffer_size(&rb), 2);
+  EXPECT_EQ(rb.wr_pos, static_cast<size_t>(0U));
+  EXPECT_EQ(ringbuffer_size(&rb), static_cast<size_t>(2U));
   
   EXPECT_EQ(ringbuffer_get(&rb, &ch), 0);
   EXPECT_EQ(ch, static_cast<uint8_t>('3'));
